ro_optimization_frequency tested and working
