+++ src/qcvv/plots/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
+from qcvv.plots.fidelity import *
 from qcvv.plots.heatmaps import *
 from qcvv.plots.scatters import *
